+++ apps/api/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@clerk/nextjs": "^6.2.0",
     "@logtail/next": "^0.1.5",
+    "@repo/analytics": "workspace:*",
     "@repo/database": "workspace:*",
     "@repo/design-system": "workspace:*",
     "@repo/env": "workspace:*",
