+++ apps/web/app/layout.tsx
@@ -1,7 +1,6 @@
 import '@repo/design-system/styles/globals.css';
 import { fonts } from '@repo/design-system/lib/fonts';
 import { DesignSystemProvider } from '@repo/design-system/providers';
-import { Analytics } from '@vercel/analytics/react';
 import type { ReactNode } from 'react';
 import { Footer } from './components/footer';
 import { Header } from './components/header';
@@ -18,7 +17,6 @@ const RootLayout = ({ children }: RootLayoutProperties) => (
         {children}
         <Footer />
       </DesignSystemProvider>
-      <Analytics />
     </body>
   </html>
 );
