+++ packages/analytics/package.json
@@ -9,5 +9,12 @@
     "posthog-js": "^1.181.0",
     "posthog-node": "^4.2.1",
     "server-only": "^0.0.1"
+  },
+  "devDependencies": {
+    "@repo/typescript-config": "workspace:*",
+    "react": "^18.3.1"
+  },
+  "peerDependencies": {
+    "react": "^18.3.1"
   }
 }
