--- packages/analytics/posthog/client.ts
@@ -1,10 +1,0 @@
-import { env } from '@repo/env';
-import posthog, { type PostHog } from 'posthog-js';
-
-export const posthogClient = posthog.init(env.NEXT_PUBLIC_POSTHOG_KEY, {
-  api_host: '/ingest',
-  ui_host: env.NEXT_PUBLIC_POSTHOG_HOST,
-  person_profiles: 'identified_only',
-  capture_pageview: false, // Disable automatic pageview capture, as we capture manually
-  capture_pageleave: true, // Overrides the `capture_pageview` setting
-}) as PostHog;
