+++ packages/analytics/posthog/server.ts
@@ -2,7 +2,7 @@ import 'server-only';
 import { env } from '@repo/env';
 import { PostHog } from 'posthog-node';
 
-export const posthogServer = new PostHog(env.NEXT_PUBLIC_POSTHOG_KEY, {
+export const posthog = new PostHog(env.NEXT_PUBLIC_POSTHOG_KEY, {
   host: env.NEXT_PUBLIC_POSTHOG_HOST,
 
   // Don't batch events and flush immediately - we're running in a serverless environment
