+++ packages/design-system/package.json
@@ -36,6 +36,7 @@
     "@radix-ui/react-toggle": "^1.1.0",
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.3",
+    "@repo/analytics": "workspace:*",
     "@repo/database": "workspace:*",
     "@repo/env": "workspace:*",
     "@sentry/nextjs": "^8.37.1",
@@ -55,8 +56,6 @@
     "lucide-react": "^0.454.0",
     "next": "15.0.2",
     "next-themes": "^0.4.3",
-    "posthog-js": "^1.181.0",
-    "posthog-node": "^4.2.1",
     "react": "18.3.1",
     "react-day-picker": "^8.10.1",
     "react-hook-form": "^7.53.1",
