+++ packages/design-system/providers/index.tsx
@@ -1,11 +1,12 @@
-import { GoogleAnalytics } from '@next/third-parties/google';
+import { GoogleAnalytics } from '@repo/analytics/google';
+import { PostHogProvider } from '@repo/analytics/posthog/client';
+import { VercelAnalytics } from '@repo/analytics/vercel';
 import { env } from '@repo/env';
-import { Analytics as VercelAnalytics } from '@vercel/analytics/react';
 import { VercelToolbar } from '@vercel/toolbar/next';
 import type { ThemeProviderProps } from 'next-themes';
 import { Toaster } from '../components/ui/sonner';
 import { TooltipProvider } from '../components/ui/tooltip';
-import { PostHogProvider } from './posthog';
+import { ClerkProvider } from './clerk';
 import { ThemeProvider } from './theme';
 
 type DesignSystemProviderProperties = ThemeProviderProps;
@@ -14,16 +15,18 @@ export const DesignSystemProvider = ({
   children,
   ...properties
 }: DesignSystemProviderProperties) => (
-  <PostHogProvider>
-    <ThemeProvider {...properties}>
-      <TooltipProvider>{children}</TooltipProvider>
-      <Toaster />
-    </ThemeProvider>
-    <VercelAnalytics />
-    {env.NODE_ENV === 'development' ? (
-      <VercelToolbar />
-    ) : (
-      <GoogleAnalytics gaId={env.NEXT_PUBLIC_GA_MEASUREMENT_ID} />
-    )}
-  </PostHogProvider>
+  <ThemeProvider {...properties}>
+    <ClerkProvider>
+      <PostHogProvider>
+        <TooltipProvider>{children}</TooltipProvider>
+        <Toaster />
+        <VercelAnalytics />
+        {env.NODE_ENV === 'development' ? (
+          <VercelToolbar />
+        ) : (
+          <GoogleAnalytics gaId={env.NEXT_PUBLIC_GA_MEASUREMENT_ID} />
+        )}
+      </PostHogProvider>
+    </ClerkProvider>
+  </ThemeProvider>
 );
